requirements.txt
absl-py==1.2.0
aiomysql==0.1.1
anyio==3.6.1
astunparse==1.6.3
bcrypt==4.0.0
cachetools==5.2.0
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cycler==0.11.0
ecdsa==0.18.0
fastapi==0.85.0
FastAPI-SQLAlchemy==0.2.1
flatbuffers==2.0.7
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.48.1
h11==0.13.0
h5py==3.7.0
httptools==0.5.0
idna==3.4
joblib==1.1.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
passlib==1.7.4
Pillow==9.2.0
plotly==5.10.0
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
starlette==0.20.4
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
threadpoolctl==3.1.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1